+++ R/f7List.R
@@ -351,3 +351,370 @@ f7ListIndex <- function(..., id) {
 #'
 #' @export
 f7ListIndexItem <- htmltools::tags$li
+
+
+
+
+#' High performance list component
+#'
+#' Use if many components in \link{f7List}
+#'
+#' @param id Virtual list unique id.
+#' @param items List items. Slot for \link{f7VirtualListItem}.
+#' @param rowsBefore Amount of rows (items) to be rendered before current
+#' screen scroll position. By default it is equal to double amount of
+#' rows (items) that fit to screen.
+#' @param rowsAfter Amount of rows (items) to be rendered after current
+#' screen scroll position. By default it is equal to the amount of rows
+#' (items) that fit to screen.
+#' @param cache Disable or enable DOM cache for already rendered list items.
+#' In this case each item will be rendered only once and all futher
+#' manipulations will be with DOM element. It is useful if your list
+#' items have some user interaction elements (like form elements or swipe outs)
+#' or could be modified.
+#'
+#' @export
+#' @examples
+#' if (interactive()) {
+#'  library(shiny)
+#'  library(shinyMobile)
+#'  shiny::shinyApp(
+#'   ui = f7Page(
+#'     title = "My app",
+#'     f7SingleLayout(
+#'       navbar = f7Navbar(
+#'         title = "Virtual Lists",
+#'         hairline = FALSE,
+#'         shadow = TRUE
+#'       ),
+#'       # main content
+#'       f7VirtualList(
+#'         id = "vlist",
+#'         rowsBefore = 2,
+#'         rowsAfter = 2,
+#'         items = lapply(1:20000, function(i) {
+#'           f7VirtualListItem(
+#'             title = paste("Title", i),
+#'             subtitle = paste("Subtitle", i),
+#'             header = paste("Header", i),
+#'             footer = paste("Footer", i),
+#'             right = paste("Right", i),
+#'             content = i,
+#'             media = img(src = "https://cdn.framework7.io/placeholder/fashion-88x88-1.jpg"),
+#'             url = NULL
+#'           )
+#'         })
+#'       )
+#'     )
+#'   ),
+#'   server = function(input, output) {
+#'
+#'   }
+#'  )
+#'
+#'  # below example will not load with classic f7List
+#'  shiny::shinyApp(
+#'    ui = f7Page(
+#'      title = "My app",
+#'      f7SingleLayout(
+#'        navbar = f7Navbar(
+#'          title = "Virtual Lists",
+#'          hairline = FALSE,
+#'          shadow = TRUE
+#'        ),
+#'        # main content
+#'        f7List(
+#'          lapply(1:20000, function(i) {
+#'            f7ListItem(
+#'              title = paste("Title", i),
+#'              subtitle = paste("Subtitle", i),
+#'              header = paste("Header", i),
+#'              footer = paste("Footer", i),
+#'              right = paste("Right", i),
+#'              content = i,
+#'              media = NULL,
+#'              url = NULL
+#'            )
+#'          })
+#'        )
+#'      )
+#'    ),
+#'    server = function(input, output) {
+#'
+#'    }
+#'  )
+#' }
+f7VirtualList <- function(id, items, rowsBefore = NULL, rowsAfter = NULL,
+                          cache = TRUE) {
+
+  config <- dropNulls(list(
+    items = items,
+    rowsBefore = rowsBefore,
+    rowsAfter = rowsAfter,
+    cache = cache
+  ))
+
+  shiny::tagList(
+    f7InputsDeps(),
+    shiny::tags$div(
+      id = id,
+      shiny::tags$script(
+        type = "application/json",
+        `data-for` = id,
+        jsonlite::toJSON(
+          x = config,
+          auto_unbox = TRUE,
+          json_verbatim = TRUE
+        )
+      ),
+      class = "list virtual-list media-list searchbar-found"
+    )
+  )
+}
+
+
+#' Virtual List item
+#'
+#' Item component for \link{f7VirtualList}
+#'
+#' @inheritParams f7ListItem
+#' @export
+f7VirtualListItem <- function(..., title = NULL, subtitle = NULL, header = NULL, footer = NULL,
+                              url = NULL, media = NULL, right = NULL) {
+
+  dropNulls(
+    list(
+      content = ...,
+      title = title,
+      subtitle = subtitle,
+      header = header,
+      footer = footer,
+      url = url,
+      media = as.character(media), # avoid issue on JS side
+      right = right
+    )
+  )
+}
+
+
+
+
+#' Update a \link{f7VirtualList} on the server side
+#'
+#' This function wraps all methods from \url{https://framework7.io/docs/virtual-list.html}
+#'
+#' @param id \link{f7VirtualList} to update.
+#' @param action Action to perform. See \url{https://framework7.io/docs/virtual-list.html}.
+#' @param item If action is one of appendItem, prependItem, replaceItem, insertItemBefore.
+#' @param items If action is one of appendItems, prependItems, replaceAllItems.
+#' @param index If action is one of replaceItem, insertItemBefore, deleteItem.
+#' @param indexes If action if one of filterItems, deleteItems.
+#' @param old_index If action is moveItem.
+#' @param new_index If action is moveItem.
+#' @param session Shiny session.
+#'
+#' @export
+#'
+#' @examples
+#' if (interactive()) {
+#'  library(shiny)
+#'  library(shinyMobile)
+#'  shiny::shinyApp(
+#'    ui = f7Page(
+#'      title = "My app",
+#'      f7SingleLayout(
+#'        navbar = f7Navbar(
+#'          title = "Virtual Lists",
+#'          hairline = FALSE,
+#'          shadow = TRUE
+#'        ),
+#'        # main content
+#'        f7Segment(
+#'          container = "segment",
+#'
+#'          f7Button(inputId = "appendItem", "Append Item"),
+#'          f7Button(inputId = "prependItems", "Prepend Items"),
+#'          f7Button(inputId = "insertBefore", "Insert before"),
+#'          f7Button(inputId = "replaceItem", "Replace Item")
+#'        ),
+#'        f7Segment(
+#'          container = "segment",
+#'          f7Button(inputId = "deleteAllItems", "Remove All"),
+#'          f7Button(inputId = "moveItem", "Move Item"),
+#'          f7Button(inputId = "filterItems", "Filter Items")
+#'        ),
+#'        f7Flex(
+#'          uiOutput("itemIndexUI"),
+#'          uiOutput("itemNewIndexUI"),
+#'          uiOutput("itemsFilterUI")
+#'        ),
+#'        f7VirtualList(
+#'          id = "vlist",
+#'          items = lapply(1:5, function(i) {
+#'            f7VirtualListItem(
+#'              title = paste("Title", i),
+#'              subtitle = paste("Subtitle", i),
+#'              header = paste("Header", i),
+#'              footer = paste("Footer", i),
+#'              right = paste("Right", i),
+#'              content = i,
+#'              media = img(src = "https://cdn.framework7.io/placeholder/fashion-88x88-3.jpg")
+#'            )
+#'          })
+#'        )
+#'      )
+#'    ),
+#'    server = function(input, output, session) {
+#'
+#'      output$itemIndexUI <- renderUI({
+#'        req(input$vlist$length > 2)
+#'        f7Stepper(
+#'          inputId = "itemIndex",
+#'          label = "Index",
+#'          min = 1,
+#'          value = 2,
+#'          max = input$vlist$length
+#'        )
+#'      })
+#'
+#'      output$itemNewIndexUI <- renderUI({
+#'        req(input$vlist$length > 2)
+#'        f7Stepper(
+#'          inputId = "itemNewIndex",
+#'          label = "New Index",
+#'          min = 1,
+#'          value = 1,
+#'          max = input$vlist$length
+#'        )
+#'      })
+#'
+#'      output$itemsFilterUI <- renderUI({
+#'        input$appendItem
+#'        input$prependItems
+#'        input$insertBefore
+#'        input$replaceItem
+#'        input$deleteAllItems
+#'        input$moveItem
+#'        isolate({
+#'          req(input$vlist$length > 2)
+#'          f7Slider(
+#'            inputId = "itemsFilter",
+#'            label = "Items to Filter",
+#'            min = 1,
+#'            max = input$vlist$length,
+#'            value = c(1, input$vlist$length)
+#'          )
+#'        })
+#'      })
+#'
+#'      observe(print(input$vlist))
+#'
+#'      observeEvent(input$appendItem, {
+#'        updateF7VirtualList(
+#'          id = "vlist",
+#'          action = "appendItem",
+#'          item = f7VirtualListItem(
+#'            title = "New Item Title",
+#'            right = "New Item Right",
+#'            content = "New Item Content",
+#'            media = img(src = "https://cdn.framework7.io/placeholder/fashion-88x88-1.jpg")
+#'          )
+#'        )
+#'      })
+#'
+#'      observeEvent(input$prependItems, {
+#'        updateF7VirtualList(
+#'          id = "vlist",
+#'          action = "prependItems",
+#'          items = lapply(1:5, function(i) {
+#'            f7VirtualListItem(
+#'              title = paste("Title", i),
+#'              right = paste("Right", i),
+#'              content = i,
+#'              media = img(src = "https://cdn.framework7.io/placeholder/fashion-88x88-1.jpg")
+#'            )
+#'          })
+#'        )
+#'      })
+#'
+#'      observeEvent(input$insertBefore, {
+#'        updateF7VirtualList(
+#'          id = "vlist",
+#'          action = "insertItemBefore",
+#'          index = input$itemIndex,
+#'          item = f7VirtualListItem(
+#'            title = "New Item Title",
+#'            content = "New Item Content",
+#'            media = img(src = "https://cdn.framework7.io/placeholder/fashion-88x88-1.jpg")
+#'          )
+#'        )
+#'      })
+#'
+#'      observeEvent(input$replaceItem, {
+#'        updateF7VirtualList(
+#'          id = "vlist",
+#'          action = "replaceItem",
+#'          index = input$itemIndex,
+#'          item = f7VirtualListItem(
+#'            title = "Replacement",
+#'            content = "Replacement Content",
+#'            media = img(src = "https://cdn.framework7.io/placeholder/fashion-88x88-1.jpg")
+#'          )
+#'        )
+#'      })
+#'
+#'      observeEvent(input$deleteAllItems, {
+#'        updateF7VirtualList(
+#'          id = "vlist",
+#'          action = "deleteAllItems"
+#'        )
+#'      })
+#'
+#'      observeEvent(input$moveItem, {
+#'        updateF7VirtualList(
+#'          id = "vlist",
+#'          action = "moveItem",
+#'          old_index = input$itemIndex,
+#'          new_index = input$itemNewIndex
+#'        )
+#'      })
+#'
+#'      observeEvent(input$filterItems, {
+#'        updateF7VirtualList(
+#'          id = "vlist",
+#'          action = "filterItems",
+#'          indexes = input$itemsFilter[1]:input$itemsFilter[2]
+#'        )
+#'      })
+#'
+#'    }
+#'  )
+#' }
+updateF7VirtualList <- function(id, action = c("appendItem", "appendItems", "prependItem",
+                                         "prependItems", "replaceItem", "replaceAllItems",
+                                         "moveItem", "insertItemBefore", "filterItems",
+                                         "deleteItem", "deleteAllItems", "scrollToItem"),
+                                item = NULL, items = NULL, index = NULL, indexes = NULL,
+                                old_index = NULL, new_index = NULL,
+                              session = shiny::getDefaultReactiveDomain()) {
+
+  # JavaScript starts from 0!
+  index <- index - 1
+  indexes <- indexes - 1
+  oldIndex <- old_index - 1
+  newIndex <- new_index - 1
+
+  message <- dropNulls(
+    list(
+      action = action,
+      item = item,
+      items = items,
+      index = index,
+      indexes = indexes,
+      oldIndex = oldIndex,
+      newIndex = newIndex
+    )
+  )
+
+  session$sendInputMessage(inputId = id, message)
+}